# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditFinding": {
        "properties": {
          "calculation": {
            "items": {
              "$ref": "#/components/schemas/CalculationStep"
            },
            "title": "Calculation",
            "type": "array"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "due_date": {
            "format": "date",
            "title": "Due Date",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/AuditEvidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "expected_amount_centimes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Amount Centimes"
          },
          "financial_impact_centimes": {
            "title": "Financial Impact Centimes",
            "type": "integer"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "finding_type": {
            "enum": [
              "THREE_WAY_MISMATCH",
              "DUPLICATE_PAYMENT",
              "PRICE_DRIFT",
              "MISSING_CREDIT_NOTE",
              "SUPPLIER_CONCENTRATION"
            ],
            "title": "Finding Type",
            "type": "string"
          },
          "observed_amount_centimes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Amount Centimes"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "recommended_action": {
            "title": "Recommended Action",
            "type": "string"
          },
          "severity": {
            "enum": [
              "CRITICAL",
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "enum": [
              "OPEN",
              "READY_FOR_APPROVAL",
              "APPROVED",
              "RESOLVED"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "supplier_id": {
            "title": "Supplier Id",
            "type": "string"
          },
          "supplier_name": {
            "title": "Supplier Name",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "finding_id",
          "finding_type",
          "severity",
          "status",
          "title",
          "summary",
          "supplier_id",
          "supplier_name",
          "financial_impact_centimes",
          "confidence",
          "recommended_action",
          "owner",
          "due_date"
        ],
        "title": "AuditFinding",
        "type": "object"
      },
...
      "FinancialRecordListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FinancialRecord"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "FinancialRecordListResponse",
        "type": "object"
      },
      "FindingDecisionRequest": {
        "properties": {
          "action": {
            "enum": [
              "PREPARE_DISPUTE",
              "APPROVE_CORRECTED_AMOUNT",
              "DISMISS"
            ],
            "title": "Action",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "action"
        ],
        "title": "FindingDecisionRequest",
        "type": "object"
      },
      "FindingDecisionResponse": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "approved_amount_centimes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved Amount Centimes"
          },
          "approved_at": {
            "format": "date-time",
            "title": "Approved At",
            "type": "string"
          },
          "approved_by": {
            "title": "Approved By",
            "type": "string"
          },
          "dispute_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispute Reference"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "enum": [
              "APPROVED",
              "RESOLVED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "finding_id",
          "status",
          "action",
          "message",
          "approved_by",
          "approved_at"
        ],
        "title": "FindingDecisionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "MIZAN Control - AI financial control, supplier intelligence, and working-capital decision support for Moroccan distribution companies.",
    "title": "MIZAN Control API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/control-tower/audit-findings": {
      "get": {
        "operationId": "list_audit_findings_api_v1_control_tower_audit_findings_get",
        "parameters": [
          {
            "example": "string",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-User-Id"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "calculation": [
                      {
                        "expression": "string",
                        "label": "Review purchasing need",
                        "result_centimes": 1
                      }
                    ],
                    "confidence": 1.0,
                    "due_date": "string",
                    "evidence": [
                      {
                        "amount_centimes": 1,
                        "document_type": "PURCHASE_ORDER",
                        "label": "Review purchasing need",
                        "quantity": 20,
                        "reference": "string",
                        "status": "SUPPORTING",
                        "unit_price_centimes": 1850
                      }
                    ],
                    "expected_amount_centimes": 1,
                    "financial_impact_centimes": 1,
                    "finding_id": "string",
                    "finding_type": "THREE_WAY_MISMATCH",
                    "observed_amount_centimes": 1,
                    "owner": "string",
                    "recommended_action": "string",
                    "severity": "CRITICAL",
                    "status": "OPEN",
                    "summary": "string",
                    "supplier_id": "string",
                    "supplier_name": "string",
                    "title": "string"
                  }
                ],
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditFinding"
                  },
                  "title": "Response List Audit Findings Api V1 Control Tower Audit Findings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit Findings",
        "tags": [
          "control-tower"
        ]
      }
    },
    "/api/v1/control-tower/audit-findings/{finding_id}": {
      "get": {
        "operationId": "get_audit_finding_api_v1_control_tower_audit_findings__finding_id__get",
        "parameters": [
          {
            "example": "string",
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-User-Id"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "calculation": [
                    {
                      "expression": "string",
                      "label": "Review purchasing need",
                      "result_centimes": 1
                    }
                  ],
                  "confidence": 1.0,
                  "due_date": "string",
                  "evidence": [
                    {
                      "amount_centimes": 1,
                      "document_type": "PURCHASE_ORDER",
                      "label": "Review purchasing need",
                      "quantity": 20,
                      "reference": "string",
                      "status": "SUPPORTING",
                      "unit_price_centimes": 1850
                    }
                  ],
                  "expected_amount_centimes": 1,
                  "financial_impact_centimes": 1,
                  "finding_id": "string",
                  "finding_type": "THREE_WAY_MISMATCH",
                  "observed_amount_centimes": 1,
                  "owner": "string",
                  "recommended_action": "string",
                  "severity": "CRITICAL",
                  "status": "OPEN",
                  "summary": "string",
                  "supplier_id": "string",
                  "supplier_name": "string",
                  "title": "string"
                },
                "schema": {
                  "$ref": "#/components/schemas/AuditFinding"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audit Finding",
        "tags": [
          "control-tower"
        ]
      }
    },
    "/api/v1/control-tower/audit-findings/{finding_id}/decide": {
      "post": {
        "operationId": "decide_audit_finding_api_v1_control_tower_audit_findings__finding_id__decide_post",
        "parameters": [
          {
            "example": "string",
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "string"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-User-Id"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-Org-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "action": "PREPARE_DISPUTE",
                "note": "string"
              },
              "schema": {
                "$ref": "#/components/schemas/FindingDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "action": "string",
                  "approved_amount_centimes": 1,
                  "approved_at": "2026-07-25T08:00:00Z",
                  "approved_by": "string",
                  "dispute_reference": "string",
                  "finding_id": "string",
                  "message": "Cooking oil may run out in four days.",
                  "status": "APPROVED"
                },
                "schema": {
                  "$ref": "#/components/schemas/FindingDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Audit Finding",
        "tags": [
          "control-tower"
        ]
      }
    },
    "/api/v1/control-tower/audit-runs": {
      "post": {
        "operationId": "run_control_audit_api_v1_control_tower_audit_runs_post",
        "parameters": [
          {
            "example": "string",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-User-Id"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "documents_analyzed": 1,
                  "findings_created": 1,
                  "model": "string",
                  "provider": "string",
                  "run_id": "string",
                  "status": "SUCCEEDED",
                  "tool_calls": [
                    {
                      "deterministic": true,
                      "duration_ms": 1,
                      "label": "Review purchasing need",
                      "name": "string",
                      "output": "string",
                      "sequence": 1,
                      "status": "SUCCEEDED"
                    }
                  ],
                  "total_impact_centimes": 1
                },
                "schema": {
                  "$ref": "#/components/schemas/AuditRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Control Audit",
        "tags": [
          "control-tower"
        ]
      }
    },
    "/api/v1/control-tower/dashboard": {
      "get": {
        "operationId": "get_dashboard_api_v1_control_tower_dashboard_get",
        "parameters": [
          {
            "example": "string",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-User-Id"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "cash_forecast": [
                    {
                      "date": "string",
                      "inflows_centimes": 1,
                      "label": "Review purchasing need",
                      "outflows_centimes": 1,
                      "projected_balance_centimes": 1
                    }
                  ],
                  "company": {
                    "active_supplier_count": 1,
                    "city": "Berrechid",
                    "legal_name": "string",
                    "name": "string",
                    "reporting_period": "string",
                    "sector": "string",
                    "warehouse_count": 1
                  },
                  "findings": [
                    {
                      "calculation": [
                        {
                          "expression": "string",
                          "label": "Review purchasing need",
                          "result_centimes": 1
                        }
                      ],
                      "confidence": 1.0,
                      "due_date": "string",
                      "evidence": [
                        {
                          "amount_centimes": 1,
                          "document_type": "PURCHASE_ORDER",
                          "label": "Review purchasing need",
                          "quantity": 20,
                          "reference": "string",
                          "status": "SUPPORTING",
                          "unit_price_centimes": 1850
                        }
                      ],
                      "expected_amount_centimes": 1,
                      "financial_impact_centimes": 1,
                      "finding_id": "string",
                      "finding_type": "THREE_WAY_MISMATCH",
                      "observed_amount_centimes": 1,
                      "owner": "string",
                      "recommended_action": "string",
                      "severity": "CRITICAL",
                      "status": "OPEN",
                      "summary": "string",
                      "supplier_id": "string",
                      "supplier_name": "string",
                      "title": "string"
                    }
                  ],
                  "generated_at": "2026-07-25T08:00:00Z",
                  "kpis": {
                    "cash_at_risk_centimes": 1,
                    "critical_findings": 1,
                    "inventory_value_centimes": 1,
                    "monitored_spend_centimes": 1,
                    "open_findings": 1,
                    "preventable_leakage_centimes": 1
                  },
                  "priority_actions": [
                    {
                      "action_id": "string",
                      "description": "string",
                      "impact_centimes": 1,
                      "target_path": "string",
                      "title": "string",
                      "urgency": "NOW"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/ControlTowerDashboard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dashboard",
        "tags": [
          "control-tower"
        ]
      }
    },
    "/api/v1/control-tower/records": {
      "get": {
        "operationId": "get_financial_records_api_v1_control_tower_records_get",
        "parameters": [
          {
            "example": "string",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-User-Id"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "items": [
                    {
                      "amount_centimes": 1,
                      "counterparty": "string",
                      "due_on": "string",
                      "issued_on": "string",
                      "linked_records": [
                        "string"
                      ],
                      "record_id": "string",
                      "record_type": "PURCHASE_ORDER",
                      "reference": "string",
                      "status": "MATCHED"
                    }
                  ]
                },
                "schema": {
                  "$ref": "#/components/schemas/FinancialRecordListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Financial Records",
        "tags": [
          "control-tower"
        ]
      }
    },
    "/api/v1/control-tower/suppliers": {
      "get": {
        "operationId": "get_supplier_portfolio_api_v1_control_tower_suppliers_get",
        "parameters": [
          {
            "example": "string",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-User-Id"
            }
          },
          {
            "example": "string",
            "in": "header",
            "name": "x-test-org-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Test-Org-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "concentration_risk_percent": 1.0,
                  "savings_opportunity_centimes": 1,
                  "scorecards": [
                    {
                      "annual_spend_centimes": 1,
                      "average_payment_terms_days": 1,
                      "category": "string",
                      "city": "Berrechid",
                      "contract_compliance_percent": 1.0,
                      "delivery_reliability_percent": 1.0,
                      "disputed_invoice_rate_percent": 1.0,
                      "name": "string",
                      "recommendation": "string",
                      "risk": "LOW",
                      "spend_share_percent": 1.0,
                      "supplier_id": "string",
                      "trend": "IMPROVING"
                    }
                  ],
                  "total_spend_centimes": 1
                },
                "schema": {
                  "$ref": "#/components/schemas/SupplierPortfolioResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Supplier Portfolio",
        "tags": [
          "control-tower"
        ]
      }
    },
...
  }
}